refactor(tests): split up mews test file
