+++ tests/src/mewsfeed/mews/all-mews.test.ts
@@ -1,0 +1,50 @@
+import { assert, test } from "vitest";
+import { runScenario, pause } from "@holochain/tryorama";
+import { Record } from "@holochain/client";
+import { createMew } from "./common";
+import { mewsfeedAppBundleSource } from "../../utils";
+
+test("create a Mew and get all mews", async () => {
+  await runScenario(
+    async (scenario) => {
+      // Set up the app to be installed
+      const appSource = { appBundleSource: mewsfeedAppBundleSource };
+
+      // Add 2 players with the test app to the Scenario. The returned players
+      // can be destructured.
+      const [alice, bob] = await scenario.addPlayersWithApps([
+        appSource,
+        appSource,
+      ]);
+
+      // Shortcut peer discovery through gossip and register all agents in every
+      // conductor of the scenario.
+      await scenario.shareAllAgents();
+
+      // Bob gets all mews
+      let collectionOutput: Record[] = await bob.cells[0].callZome({
+        zome_name: "mews",
+        fn_name: "get_all_mews",
+        payload: null,
+      });
+      assert.equal(collectionOutput.length, 0);
+
+      // Alice creates a Mew
+      const createdRecord: Record = await createMew(alice.cells[0]);
+      assert.ok(createdRecord);
+
+      await pause(1200);
+
+      // Bob gets all mews again
+      collectionOutput = await bob.cells[0].callZome({
+        zome_name: "mews",
+        fn_name: "get_all_mews",
+        payload: null,
+      });
+      assert.equal(collectionOutput.length, 1);
+      assert.deepEqual(createdRecord, collectionOutput[0]);
+    },
+    true,
+    { timeout: 100000 }
+  );
+});
