+++ tests/vitest.config.ts
@@ -4,5 +4,7 @@ export default defineConfig({
   test: {
     threads: false,
     testTimeout: 60 * 1000 * 3, // 3  mins
+    outputDiffLines: 30,
+    outputDiffMaxSize: 30000,
   },
 });
